+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 3.4 (not yet released)
 
+* Fixed issue #220: The condition "contact custom field changed is one of" does not work when it is a date field.
 * Fix crash when contact_id is array / undefined on post trigger (eg. for Case entity).
 
 ## Version 3.3 (2024-04-06)